后台管理模块：完成品牌管理curd

+++ spzx-manager/src/main/java/com/yinxin/spzx/manager/service/impl/SysUserServiceImpl.java
@@ -93,7 +93,9 @@ public class SysUserServiceImpl implements SysUserService {
 
         SysUser user = sysUserMapper.getById(sysUser.getId());
         sysUserMapper.update(sysUser);
-        MinIoUtil.removeFile(user.getAvatar());
+        if (!StrUtil.isEmpty(sysUser.getAvatar())) {
+            MinIoUtil.removeFile(user.getAvatar());
+        }
     }
 
     @Override
